# swagger.yaml
...
swagger: '2.0'
info:
  title: Herr Ober API
  description: Api fuer Herr Ober
  version: "1.0.0"
# the domain of the service
host: herrober.rdeadmin.waters.com
...
schemes:
  - https
# will be prefixed to all paths
basePath: /ober
produces:
  - application/json
paths:
...
  /orders:
    get:
      summary: orders
      description: the orders
      tags:
        - orders
      parameters:
        - name: date
          in: query
          type: string
          description: delivery date and time
        - name: orderstatus
          in: query
          type: string
          description: whether the order is open on its way or already finished
          enum:
            - open
            - ordered
            - delivered
      responses:
        200:
          description: the current amount of money
          schema:
            type: array
            items:
              $ref: '#/definitions/order'
    put:
      tags:
        - orders
      summary: add or update an order
      parameters:
        - in: body
          name: body
          description: order to add or update
          required: true
          schema:
            $ref: '#/definitions/order'
      responses:
        204:
          description: 'Request successfull with no response content.'
    delete:
      tags:
        - orders
      summary: delete an order
      parameters:
        - name: userEmail
          in: query
          required: true
          type: string
          description: Unique identifier of a user
        - name: plannedDateOfDelivery
          in: query
          required: true
          type: string
          description: delivery date and time
        - name: restaurant
          in: query
          required: true
          type: string
          description: name of the restaurant
        - name: dishName
          in: query
          required: true
          type: string
          description: name of the dish
      responses:
        204:
          description: 'Ok, first matching order delete'
        404:
          description: 'Not found a matching order'
        406:
          description: 'incorrect parameters'
        
    patch:
      tags:
        - orders
      summary: pay a delivery
      parameters:
        - name: user
          in: query
          required: true
          type: string
          description: Unique identifier of a user
        - name: date
          in: query
          required: true
          type: string
          description: delivery date and time
        - name: restaurant
          in: query
          required: true
          type: string
          description: name of the restaurant
        - name: amount
          in: query
          type: number
          description: amount that the current user has paid.
      responses:
        204:
          description: 'Request successfull with no response content.'
...
definitions:
...
  order:
    type: object
    properties:
      restaurant:
        type: string
        description: name
      plannedDeliveryDate:
        type: string
        description: planned delivery date and time
      email:
        type: string
        description: email address
      plateId:
        type: string
        description: identifier of the dish
      baseprice:
        type: number
        description: identifier of the dish
      extras:
        type: string
        description: extras for the current product (e.g. for a Pizza)
      extrasprice:
        type: number
        description: price for the extras
      orderstatus:
          type: string
          description: whether the order is open on its way or already finished
          enum:
            - open
            - ordered
            - delivered
...